+++ lib/src/presentation/widgets/questionnaire_item/display/questionnaire_display_item_view.dart
@@ -29,7 +29,7 @@ class QuestionnaireDisplayItemViewState
       super.buildTitleView(
         context,
         noPadding: true,
-        style: theme.textTheme.titleMedium,
+        style: theme.textTheme.bodyMedium,
       ) ??
       SizedBox();
 }